+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv (development version)
 
+* Fixed an issue where package installation could fail on Arch Linux with the
+  error message "don't know how to check sysreqs on this system". (#2107)
+
 * `renv` now supports profile-specific comments within `.gitignore` and
   `.renvignore` files. These are comments of the form `#| <code>`, where
   `<code>` is R code which should indicate if the `.renvignore` rules should